requirements.txt
google-genai
sounddevice
numpy
scipy
keyboard
pyperclip
python-dotenv
Pillow
pyautogui
pygetwindow
